+++ packages/block/test/block.spec.ts
@@ -1,13 +1,9 @@
 import tape from 'tape'
-import { keccak256 } from 'ethereum-cryptography/keccak'
-import { bytesToHex } from 'ethereum-cryptography/utils'
-import { bufArrToArr, NestedUint8Array, toBuffer, zeros } from '@ethereumjs/util'
+import { NestedUint8Array, toBuffer, zeros } from '@ethereumjs/util'
 import RLP from 'rlp'
 import Common, { Chain, Hardfork } from '@ethereumjs/common'
-import { Block, BlockBuffer, BlockHeader } from '../src'
+import { Block, BlockBuffer } from '../src'
 import blockFromRpc from '../src/from-rpc'
-import { Mockchain } from './mockchain'
-import { createBlock } from './util'
 import testnetMerge from './testdata/testnetMerge.json'
 import * as testDataPreLondon from './testdata/testdata_pre-london.json'
 import * as testDataPreLondon2 from './testdata/testdata_pre-london-2.json'
@@ -232,388 +228,8 @@ tape('[Block]: block functions', function (t) {
       st.ok(error.message.includes('invalid uncle hash'))
     }
   })
-  /*
-  t.test('should throw if an uncle is listed twice', async function (st) {
-    const common = new Common({ chain: Chain.Mainnet, hardfork: Hardfork.Chainstart })
-    const blockchain = new Mockchain()
 
-    const genesis = Block.fromBlockData({})
-    await blockchain.putBlock(genesis)
-
-    const uncleBlock1 = createBlock(genesis, 'uncle', [], common)
-
-    const block1 = createBlock(genesis, 'block1', [], common)
-    const block2 = createBlock(block1, 'block1', [uncleBlock1.header, uncleBlock1.header], common)
-
-    await blockchain.putBlock(uncleBlock1)
-    await blockchain.putBlock(block1)
-    await blockchain.putBlock(block2)
-
-    try {
-      await block2.validate(blockchain)
-      st.fail('cannot reach this')
-    } catch (e: any) {
-      st.pass('block throws if the uncle is included twice in the block')
-    }
-  })
-
-  t.test('should throw if an uncle is included before', async function (st) {
-    const common = new Common({ chain: Chain.Mainnet, hardfork: Hardfork.Chainstart })
-    const blockchain = new Mockchain()
-
-    const genesis = Block.fromBlockData({})
-    await blockchain.putBlock(genesis)
-
-    const uncleBlock = createBlock(genesis, 'uncle', [], common)
-
-    const block1 = createBlock(genesis, 'block1', [], common)
-    const block2 = createBlock(block1, 'block2', [uncleBlock.header], common)
-    const block3 = createBlock(block2, 'block3', [uncleBlock.header], common)
-
-    await blockchain.putBlock(uncleBlock)
-    await blockchain.putBlock(block1)
-    await blockchain.putBlock(block2)
-    await blockchain.putBlock(block3)
-
-    await uncleBlock.validate(blockchain)
-
-    await block1.validate(blockchain)
-    await block2.validate(blockchain)
-
-    try {
-      await block3.validate(blockchain)
-      st.fail('cannot reach this')
-    } catch (e: any) {
-      st.pass('block throws if uncle is already included')
-    }
-  })
-
-  t.test(
-    'should throw if the uncle parent block is not part of the canonical chain',
-    async function (st) {
-      const common = new Common({ chain: Chain.Mainnet, hardfork: Hardfork.Chainstart })
-      const blockchain = new Mockchain()
-
-      const genesis = Block.fromBlockData({})
-      await blockchain.putBlock(genesis)
-
-      const emptyBlock = Block.fromBlockData({ header: { number: BigInt(1) } }, { common })
-
-      //assertion
-      if (emptyBlock.hash().equals(genesis.hash())) {
-        st.fail('should create an unique bogus block')
-      }
-
-      await blockchain.putBlock(emptyBlock)
-
-      const uncleBlock = createBlock(emptyBlock, 'uncle', [], common)
-      const block1 = createBlock(genesis, 'block1', [], common)
-      const block2 = createBlock(block1, 'block2', [], common)
-      const block3 = createBlock(block2, 'block3', [uncleBlock.header], common)
-
-      await blockchain.putBlock(uncleBlock)
-      await blockchain.putBlock(block1)
-      await blockchain.putBlock(block2)
-      await blockchain.putBlock(block3)
-
-      try {
-        await block3.validate(blockchain)
-        st.fail('cannot reach this')
-      } catch (e: any) {
-        st.pass('block throws if uncle parent hash is not part of the canonical chain')
-      }
-    }
-  )
-
-  t.test('should throw if the uncle is too old', async function (st) {
-    const common = new Common({ chain: Chain.Mainnet, hardfork: Hardfork.Chainstart })
-    const blockchain = new Mockchain()
-
-    const genesis = Block.fromBlockData({})
-    await blockchain.putBlock(genesis)
-
-    const uncleBlock = createBlock(genesis, 'uncle', [], common)
-
-    let lastBlock = genesis
-    for (let i = 0; i < 7; i++) {
-      const block = createBlock(lastBlock, 'block' + i.toString(), [], common)
-      await blockchain.putBlock(block)
-      lastBlock = block
-    }
-
-    const blockWithUnclesTooOld = createBlock(
-      lastBlock,
-      'too-old-uncle',
-      [uncleBlock.header],
-      common
-    )
-
-    try {
-      await blockWithUnclesTooOld.validate(blockchain)
-      st.fail('cannot reach this')
-    } catch (e: any) {
-      st.pass('block throws uncle is too old')
-    }
-  })
-
-  t.test('should throw if uncle is too young', async function (st) {
-    const common = new Common({ chain: Chain.Mainnet, hardfork: Hardfork.Chainstart })
-    const blockchain = new Mockchain()
-
-    const genesis = Block.fromBlockData({})
-    await blockchain.putBlock(genesis)
-
-    const uncleBlock = createBlock(genesis, 'uncle', [], common)
-    const block1 = createBlock(genesis, 'block1', [uncleBlock.header], common)
-
-    await blockchain.putBlock(uncleBlock)
-    await blockchain.putBlock(block1)
-
-    try {
-      await block1.validate(blockchain)
-      st.fail('cannot reach this')
-    } catch (e: any) {
-      st.pass('block throws uncle is too young')
-    }
-  })
-
-  t.test('should throw if the uncle header is invalid', async function (st) {
-    const common = new Common({ chain: Chain.Mainnet, hardfork: Hardfork.Chainstart })
-    const blockchain = new Mockchain()
-
-    const genesis = Block.fromBlockData({})
-    await blockchain.putBlock(genesis)
-
-    const uncleBlock = Block.fromBlockData(
-      {
-        header: {
-          number: genesis.header.number + BigInt(1),
-          parentHash: genesis.hash(),
-          timestamp: genesis.header.timestamp + BigInt(1),
-          gasLimit: BigInt(5000),
-          difficulty: BigInt(0), // invalid difficulty
-        },
-      },
-      { common }
-    )
-
-    const block1 = createBlock(genesis, 'block1', [], common)
-    const block2 = createBlock(block1, 'block2', [uncleBlock.header], common)
-
-    await blockchain.putBlock(uncleBlock)
-    await blockchain.putBlock(block1)
-    await blockchain.putBlock(block2)
-
-    try {
-      await block2.validate(blockchain)
-      st.fail('cannot reach this')
-    } catch (e: any) {
-      st.pass('block throws uncle header is invalid')
-    }
-  })
-
-  t.test('throws if more than 2 uncles included', async function (st) {
-    const common = new Common({ chain: Chain.Mainnet, hardfork: Hardfork.Chainstart })
-    const blockchain = new Mockchain()
-
-    const genesis = Block.fromBlockData({})
-    await blockchain.putBlock(genesis)
-
-    const uncleBlock1 = createBlock(genesis, 'uncle1', [], common)
-    const uncleBlock2 = createBlock(genesis, 'uncle2', [], common)
-    const uncleBlock3 = createBlock(genesis, 'uncle3', [], common)
-
-    // sanity check
-    if (
-      uncleBlock1.hash().equals(uncleBlock2.hash()) ||
-      uncleBlock2.hash().equals(uncleBlock3.hash())
-    ) {
-      st.fail('uncles 1/2/3 should be unique')
-    }
-
-    const block1 = createBlock(genesis, 'block1', [], common)
-    const block2 = createBlock(
-      block1,
-      'block1',
-      [uncleBlock1.header, uncleBlock2.header, uncleBlock3.header],
-      common
-    )
-
-    await blockchain.putBlock(uncleBlock1)
-    await blockchain.putBlock(uncleBlock2)
-    await blockchain.putBlock(uncleBlock3)
-    await blockchain.putBlock(block1)
-    await blockchain.putBlock(block2)
-
-    try {
-      await block2.validate(blockchain)
-      st.fail('cannot reach this')
-    } catch (e: any) {
-      st.pass('block throws if more than 2 uncles are included')
-    }
-  })
-
-  t.test('throws if uncle is a canonical block', async function (st) {
-    const common = new Common({ chain: Chain.Mainnet, hardfork: Hardfork.Chainstart })
-    const blockchain = new Mockchain()
-
-    const genesis = Block.fromBlockData({})
-    await blockchain.putBlock(genesis)
-
-    const block1 = createBlock(genesis, 'block1', [], common)
-    const block2 = createBlock(block1, 'block2', [block1.header], common)
-
-    await blockchain.putBlock(block1)
-    await blockchain.putBlock(block2)
-
-    try {
-      await block2.validate(blockchain)
-      st.fail('cannot reach this')
-    } catch (e: any) {
-      st.pass('block throws if an uncle is a canonical block')
-    }
-  })
-
-  t.test('successfully validates uncles', async function (st) {
-    const common = new Common({ chain: Chain.Mainnet, hardfork: Hardfork.Chainstart })
-    const blockchain = new Mockchain()
-
-    const genesis = Block.fromBlockData({})
-    await blockchain.putBlock(genesis)
-
-    const uncleBlock = createBlock(genesis, 'uncle', [], common)
-    await blockchain.putBlock(uncleBlock)
-
-    const block1 = createBlock(genesis, 'block1', [], common)
-    const block2 = createBlock(block1, 'block2', [uncleBlock.header], common)
-
-    await blockchain.putBlock(block1)
-    await blockchain.putBlock(block2)
-
-    await block1.validate(blockchain)
-    await block2.validate(blockchain)
-    st.pass('uncle blocks validated succesfully')
-  })
-*/
-  /*  t.test(
-    'should select the right hardfork for uncles at a hardfork transition',
-    async function (st) {
- */ /**
-   * This test creates a chain around mainnet fork blocks:
-   *      berlin         london
-   *                |     |-> u <---|
-   * @ -> @ -> @ ---|---> @ -> @ -> @
-   * |-> u <---|               | -> @
-   *    ^----------------------------
-   * @ = block
-   * u = uncle block
-   *
-   * There are 3 pre-fork blocks, with 1 pre-fork uncle
-   * There are 3 blocks after the fork, with 1 uncle after the fork
-   *
-   * The following situations are tested:
-   * Pre-fork block can have legacy uncles
-   * London block has london uncles
-   * London block has legacy uncles
-   * London block has legacy uncles, where hardforkByBlockNumber set to false (this should throw)
-   *    In this situation, the london block creates a london uncle, but this london uncle should be
-   *    a berlin block, and therefore has no base fee. But, since common is still london, base fee
-   *    is expected
-   * It is tested that common does not change
-   */ /*     const blockchain = new Mockchain()
-
-      const common = new Common({ chain: Chain.Mainnet })
-      common.setHardfork(Hardfork.Berlin)
-
-      const mainnetForkBlock = common.hardforkBlock(Hardfork.London)
-      const rootBlock = Block.fromBlockData(
-        {
-          header: {
-            number: mainnetForkBlock! - BigInt(3),
-            gasLimit: BigInt(5000),
-          },
-        },
-        { common }
-      )
-
-      await blockchain.putBlock(rootBlock)
-
-      const unclePreFork = createBlock(rootBlock, 'unclePreFork', [], common)
-      const canonicalBlock = createBlock(rootBlock, 'canonicalBlock', [], common)
-      await blockchain.putBlock(canonicalBlock)
-      const preForkBlock = createBlock(
-        canonicalBlock,
-        'preForkBlock',
-        [unclePreFork.header],
-        common
-      )
-      await blockchain.putBlock(preForkBlock)
-      common.setHardfork(Hardfork.London)
-      const forkBlock = createBlock(preForkBlock, 'forkBlock', [], common)
-      await blockchain.putBlock(forkBlock)
-      const uncleFork = createBlock(forkBlock, 'uncleFork', [], common)
-      const canonicalBlock2 = createBlock(forkBlock, 'canonicalBlock2', [], common)
-      const forkBlock2 = createBlock(canonicalBlock2, 'forkBlock2', [uncleFork.header], common)
-      await blockchain.putBlock(canonicalBlock2)
-      await blockchain.putBlock(forkBlock)
-      await preForkBlock.validate(blockchain)
-
-      st.equal(common.hardfork(), Hardfork.London, 'validation did not change common hardfork')
-      await forkBlock2.validate(blockchain)
-
-      st.equal(common.hardfork(), Hardfork.London, 'validation did not change common hardfork')
-
-      const forkBlock2HeaderData = forkBlock2.header.toJSON()
-      const uncleHeaderData = unclePreFork.header.toJSON()
-
-      uncleHeaderData.extraData = '0xffff'
-      const uncleHeader = BlockHeader.fromHeaderData(uncleHeaderData, {
-        common: new Common({ chain: Chain.Mainnet, hardfork: Hardfork.Berlin }),
-      })
-
-      forkBlock2HeaderData.uncleHash =
-        '0x' + bytesToHex(keccak256(RLP.encode(bufArrToArr([uncleHeader.raw()]))))
-
-      const forkBlock_ValidCommon = Block.fromBlockData(
-        {
-          header: forkBlock2HeaderData,
-          uncleHeaders: [uncleHeaderData],
-        },
-        {
-          common,
-        }
-      )
-
-      await forkBlock_ValidCommon.validate(blockchain)
-
-      st.pass('successfully validated a pre-london uncle on a london block')
-      st.equal(common.hardfork(), Hardfork.London, 'validation did not change common hardfork')
-
-      const forkBlock_InvalidCommon = Block.fromBlockData(
-        {
-          header: forkBlock2HeaderData,
-          uncleHeaders: [uncleHeaderData],
-        },
-        {
-          common,
-          hardforkByBlockNumber: false,
-        }
-      )
-
-      try {
-        await forkBlock_InvalidCommon.validate(blockchain)
-        st.fail('cannot reach this')
-      } catch (e: any) {
-        st.ok(
-          e.message.includes('with EIP1559 being activated'),
-          'explicitly set hardforkByBlockNumber to false, pre-london block interpreted as london block and succesfully failed'
-        )
-      }
-
-      st.equal(common.hardfork(), Hardfork.London, 'validation did not change common hardfork')
-    }
-  )
-*/ t.test('should test isGenesis (mainnet default)', function (st) {
+  t.test('should test isGenesis (mainnet default)', function (st) {
     const block = Block.fromBlockData({ header: { number: 1 } })
     st.notEqual(block.isGenesis(), true)
     const genesisBlock = Block.fromBlockData({ header: { number: 0 } })
